# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
  model.FlagEval:
    properties:
      value: {}
    type: object
...
  model.RelayProxyRequest:
    properties:
      defaultValue: {}
      user:
        $ref: '#/definitions/model.UserRequest'
    type: object
  model.UserRequest:
    properties:
      anonymous:
        description: Anonymous set if this is a logged-in user or not.
        type: boolean
      custom:
        additionalProperties: true
        description: Custom is a map containing all extra information for this user.
        type: object
      key:
        description: Key is the identifier of the UserRequest.
        type: string
    type: object
...
info:
  contact:
    name: GO feature flag relay proxy
    url: https://github.com/thomaspoignant/go-feature-flag-relay-proxy
  description: |-
    Swagger for the go-feature-flag relay proxy.

    go-feature-flag relay proxy is using thomaspoignant/go-feature-flag to handle your feature flagging.
    It is a proxy to your flags, you can get the values of your flags using APIs.
  license:
    name: MIT
    url: https://github.com/thomaspoignant/go-feature-flag-relay-proxy/blob/main/LICENSE
  title: go-feature-flag relay proxy
paths:
...
  /v1/feature/{flag_key}/eval:
    post:
      consumes:
      - application/json
      description: allflags returns all the flag for a specific user.
      parameters:
      - description: Payload of the user we want to get all the flags from.
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.RelayProxyRequest'
      - description: Name of your feature flag
        in: path
        name: flag_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.FlagEval'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: allflags returns all the flag for a specific user.
      tags:
      - flags
swagger: "2.0"